less verbose, and guarding against clobbering

+++ ngs/sort-fastq.sh
@@ -3,7 +3,7 @@
 # stolen from dariober, https://www.biostars.org/p/15011/#103041
 
 if [ $# -ne 1 ] ;then
-  echo "Usage: $0 FILE.gz # FILE.gz is renamed FILE-unsorted.gz, sorted output to FILE.gz" >&2 
+  echo "Usage: $0 FILE.gz # FILE.gz is renamed FILE-unsorted.gz (and left there), sorted output to FILE.gz" >&2 
   exit 2
 fi
 
@@ -13,8 +13,13 @@ file=$1
 name=${file/.gz}
 unsorted="$name-unsorted.gz"
 
+if [  -f $unsorted ] ; then
+  echo "Not overwriting existing file $unsorted; nothing sorted." >&2 
+  exit 5
+fi 
+
 if mv $file $unsorted ; then
-  echo "Renamed $file to $unsorted" >&2   
+: #  echo "Renamed $file to $unsorted" >&2   
 else
   echo "Could not mv $file $unsorted" >&2 
   exit 3
